+++ .cargo/config.toml
@@ -3,11 +3,11 @@ tt = "nextest run --tests --no-fail-fast"                  # All
 tl = "nextest run --lib --no-fail-fast"                    # Lib only
 lx = "run --bin lx --"
 xtask = "run --package xtask --"
-xtask51 = "run --package xtask --features lua51 --"
-xtask52 = "run --package xtask --features lua52 --"
-xtask53 = "run --package xtask --features lua53 --"
-xtask54 = "run --package xtask --features lua54 --"
-xtaskjit = "run --package xtask --features luajit --"
+xtask-lua51 = "run --package xtask-lua --features lua51 --"
+xtask-lua52 = "run --package xtask-lua --features lua52 --"
+xtask-lua53 = "run --package xtask-lua --features lua53 --"
+xtask-lua54 = "run --package xtask-lua --features lua54 --"
+xtask-luajit = "run --package xtask-lua --features luajit --"
 build-all = "run --package xtask -- build"
 build-all-release = "run --package xtask -- build-release"
 dist = "run --package xtask -- dist"
