+++ lux-cli/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lux-cli"
-version = "0.4.1"
+version = "0.4.2"
 description = "A luxurious package manager for Lua"
 authors = ["vhyrro <vhyrro@gmail.com>", "mrcjkb <marc@jakobi.dev>"]
 homepage = "https://github.com/nvim-neorocks/lux"
