+++ lux-lua/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lux-lua"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 publish = false
 
@@ -21,5 +21,6 @@ lua51 = ["mlua/module", "mlua/lua51", "lux-lib/lua51"]
 lua52 = ["mlua/module", "mlua/lua52", "lux-lib/lua52"]
 lua53 = ["mlua/module", "mlua/lua53", "lux-lib/lua53"]
 lua54 = ["mlua/module", "mlua/lua54", "lux-lib/lua54"]
+luajit = ["mlua/module", "mlua/luajit", "lux-lib/luajit"]
 # Run tests without module mode so we can link against Lua and run the tests.
 test = ["mlua/vendored", "lux-lib/lua51"]
