+++ lux-workspace-hack/Cargo.toml
@@ -46,6 +46,7 @@ smallvec = { version = "1", default-features = false, features = ["const_new"] }
 time = { version = "0.3", default-features = false, features = ["formatting", "macros", "parsing"] }
 tokio = { version = "1", features = ["full"] }
 tokio-util = { version = "0.7", features = ["codec", "io"] }
+toml_edit = { version = "0.22", features = ["serde"] }
 tracing = { version = "0.1", features = ["log"] }
 
 [build-dependencies]
